ReactionsByTimeWidget
